+++ packages/plugins/Pins/src/store/index.ts
@@ -56,7 +56,7 @@ export const makeStoreModule = () => {
               // these indicate other interactions are expected now
               interaction instanceof Draw ||
               interaction instanceof Modify ||
-              // @ts-expect-error | internal hack to detect it from Draw plugin
+              // @ts-expect-error | internal hack to detect it from @polar/plugin-draw
               interaction._isDeleteSelect
           )
           if (