+++ packages/core/src/index.ts
@@ -8,6 +8,7 @@ import './monkeyCrossOrigin'
 import './monkeyHeaderLoader'
 
 export { NineLayout, NineLayoutTag } from './components/layouts'
+export { RadioCard } from './components'
 export { setLayout, getLayout } from './utils/layout'
 export { default as mpapiDefaults } from './utils/createMap/defaults'
 export type { MapInstance } from './types'