+++ packages/plugins/Draw/src/language.ts
@@ -9,7 +9,9 @@ const language: LanguageOption[] = [
           mode: {
             none: 'Keine Interaktion',
             draw: 'Zeichnen',
+            measure: 'Zeichnen und Messen',
             write: 'Zeichnen und Schreiben',
+            writeAndMeasure: 'Zeichnen, Schreiben und Messen',
             edit: 'Bearbeiten',
             delete: 'Löschen',
           },
@@ -20,12 +22,21 @@ const language: LanguageOption[] = [
             polygon: 'Polygon',
             text: 'Text',
           },
+          measureMode: {
+            none: 'Keine Messung',
+            metres: 'm',
+            metresArea: 'm / m²',
+            kilometres: 'km',
+            kilometresArea: 'km / km²',
+            hectares: 'km / ha',
+          },
           options: {
             stroke: 'Linienfarbe',
           },
           title: {
-            mode: 'Modus',
             drawMode: 'Zeichenmodus',
+            mode: 'Modus',
+            measureMode: 'Messmodus',
             options: 'Zeichenoptionen',
           },
           label: {
@@ -43,7 +54,9 @@ const language: LanguageOption[] = [
           mode: {
             none: 'No interaction',
             draw: 'Draw',
+            measure: 'Draw and measure',
             write: 'Draw and write',
+            writeAndMeasure: 'Draw, write and measure',
             edit: 'Edit',
             delete: 'Delete',
           },
@@ -54,12 +67,21 @@ const language: LanguageOption[] = [
             polygon: 'Polygon',
             text: 'Text',
           },
+          measureMode: {
+            none: 'No measure',
+            metres: 'm',
+            metresArea: 'm / m²',
+            kilometres: 'km',
+            kilometresArea: 'km / km²',
+            hectares: 'km / ha',
+          },
           options: {
             stroke: 'Stroke color',
           },
           title: {
-            mode: 'Mode',
             drawMode: 'Drawing mode',
+            mode: 'Mode',
+            measureMode: 'Measure mode',
             options: 'Draw options',
           },
           label: {