+++ packages/core/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## unpublished
 
+- Feature: Add new reusable component `RadioCard.vue` to the package.
 - Feature: Additionally export `MapInstance` type.
+- Fix: Do not break themes of external Vuetify apps.
+
+## 2.0.1
 - Fix: Add `crossOrigin` differently to layer sources that are an instance of `ImageWMS` as they require it being set as `crossOrigin_` to be recognized.
 - Fix: Add missing `font-family` css so that tooltips are always `Arial, sans-serif`.